Fix #7781: Avoid using nested package definition
Had to remove `sealed` from `Expr`
